# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "learnwithai",
    "description": "HTTP API for LearnWithAI, including operational endpoints, UNC authentication flows, and authenticated user profile access.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/courses": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "List my courses",
        "description": "Returns courses the authenticated subject is enrolled in.\n\nArgs:\n    subject: Authenticated subject.\n    course_svc: Service used to query courses.\n\nReturns:\n    List of courses with active membership.",
        "operationId": "list_my_courses",
        "responses": {
          "200": {
            "description": "Courses where the current user has active membership.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseResponse"
                  },
                  "type": "array",
                  "title": "Response List My Courses"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Create a course",
        "description": "Creates a new course and enrolls the caller as instructor.\n\nArgs:\n    subject: Authenticated subject.\n    body: Course creation payload.\n    course_svc: Service used to create the course.\n\nReturns:\n    The newly created course.",
        "operationId": "create_course",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The newly created course.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/courses/{course_id}": {
      "put": {
        "tags": [
          "Courses"
        ],
        "summary": "Update a course",
        "description": "Updates an existing course's details.\n\nOnly the instructor of the course may update it.\n\nArgs:\n    subject: Authenticated subject.\n    course: Course loaded via DI and course_id path param.\n    body: Course update payload.\n    course_svc: Service used to update the course.\n\nReturns:\n    The updated course with the caller's membership.",
        "operationId": "update_course",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated course.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Only instructors can update courses."
          },
          "404": {
            "description": "Course not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/roster": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Get course roster",
        "description": "Returns the paginated roster for a course.\n\nOnly instructors and TAs may view the roster.\n\nArgs:\n    subject: Authenticated subject.\n    course: Course loaded via DI and course_id Path param.\n    course_svc: Service used to query the roster.\n    pagination: Pagination parameters from query string.\n    q: Optional search query to filter by name, PID, or email.\n\nReturns:\n    Paginated roster with user details.",
        "operationId": "get_course_roster",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated roster with user details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRosterResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Insufficient permissions."
          },
          "404": {
            "description": "Course not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/members": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Add a member to a course",
        "description": "Adds a member to a course.\n\nOnly the instructor of the course may add members.\n\nArgs:\n    subject: Authenticated subject.\n    course: Course loaded via DI and course_id path param.\n    add_member_request: Member addition payload.\n    course_svc: Service used to manage memberships.\n    user_repo: Repository used to load the target user from the request body.\n\nReturns:\n    The newly created membership.",
        "operationId": "add_member",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created membership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Only instructors can add members."
          },
          "404": {
            "description": "Course or user not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/members/{pid}": {
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Drop a member from a course",
        "description": "Drops a member from a course.\n\nInstructors can drop anyone. Students and TAs can drop themselves.\n\nArgs:\n    subject: Authenticated subject.\n    course: Course loaded via DI and course_id path param.\n    target_user: User loaded via DI and pid path param.\n    course_svc: Service used to manage memberships.\n\nReturns:\n    The updated membership.",
        "operationId": "drop_member",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated membership with dropped state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Insufficient permissions to drop this member."
          },
          "404": {
            "description": "Course or user not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddMemberRequest": {
        "properties": {
          "pid": {
            "type": "integer",
            "title": "Pid"
          },
          "type": {
            "$ref": "#/components/schemas/MembershipType"
          }
        },
        "type": "object",
        "required": [
          "pid",
          "type"
        ],
        "title": "AddMemberRequest",
        "description": "Payload for adding a member to a course."
      },
...
      "CourseMembership": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/MembershipType"
          },
          "state": {
            "$ref": "#/components/schemas/MembershipState"
          }
        },
        "type": "object",
        "required": [
          "type",
          "state"
        ],
        "title": "CourseMembership",
        "description": "Represents the caller's membership for a returned course."
      },
      "CourseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_number": {
            "type": "string",
            "title": "Course Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "term": {
            "$ref": "#/components/schemas/Term"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "membership": {
            "$ref": "#/components/schemas/CourseMembership"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_number",
          "name",
          "description",
          "term",
          "year",
          "membership"
        ],
        "title": "CourseResponse",
        "description": "Represents a course returned by the API."
      },
      "CreateCourseRequest": {
        "properties": {
          "course_number": {
            "type": "string",
            "title": "Course Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "term": {
            "$ref": "#/components/schemas/Term"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          }
        },
        "type": "object",
        "required": [
          "course_number",
          "name",
          "term",
          "year"
        ],
        "title": "CreateCourseRequest",
        "description": "Payload for creating a new course."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MembershipResponse": {
        "properties": {
          "user_pid": {
            "type": "integer",
            "title": "User Pid"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "type": {
            "$ref": "#/components/schemas/MembershipType"
          },
          "state": {
            "$ref": "#/components/schemas/MembershipState"
          }
        },
        "type": "object",
        "required": [
          "user_pid",
          "course_id",
          "type",
          "state"
        ],
        "title": "MembershipResponse",
        "description": "Represents a membership returned by the API."
      },
      "MembershipState": {
        "type": "string",
        "enum": [
          "pending",
          "enrolled",
          "dropped"
        ],
        "title": "MembershipState",
        "description": "Lifecycle state of a course membership."
      },
      "MembershipType": {
        "type": "string",
        "enum": [
          "instructor",
          "ta",
          "student"
        ],
        "title": "MembershipType",
        "description": "Role a user holds within a course."
      },
      "PaginatedRosterResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RosterMemberResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedRosterResponse",
        "description": "Paginated wrapper around roster member entries."
      },
      "RosterMemberResponse": {
        "properties": {
          "user_pid": {
            "type": "integer",
            "title": "User Pid"
          },
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "type": {
            "$ref": "#/components/schemas/MembershipType"
          },
          "state": {
            "$ref": "#/components/schemas/MembershipState"
          },
          "given_name": {
            "type": "string",
            "title": "Given Name"
          },
          "family_name": {
            "type": "string",
            "title": "Family Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "user_pid",
          "course_id",
          "type",
          "state",
          "given_name",
          "family_name",
          "email"
        ],
        "title": "RosterMemberResponse",
        "description": "A roster entry combining membership and user details."
      },
...
      "Term": {
        "type": "string",
        "enum": [
          "fall",
          "winter",
          "spring",
          "summer"
        ],
        "title": "Term",
        "description": "Academic term within a year."
      },
      "UpdateCourseRequest": {
        "properties": {
          "course_number": {
            "type": "string",
            "title": "Course Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "term": {
            "$ref": "#/components/schemas/Term"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          }
        },
        "type": "object",
        "required": [
          "course_number",
          "name",
          "term",
          "year"
        ],
        "title": "UpdateCourseRequest",
        "description": "Payload for updating an existing course."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}